+++ modules/mlflow/mlflow-image/tests/test_stack.py
@@ -1,0 +1,43 @@
+import os
+import sys
+
+import aws_cdk as cdk
+import pytest
+from aws_cdk.assertions import Template
+
+
+@pytest.fixture(scope="function")
+def stack_defaults() -> None:
+    os.environ["CDK_DEFAULT_ACCOUNT"] = "111111111111"
+    os.environ["CDK_DEFAULT_REGION"] = "us-east-1"
+
+    # Unload the app import so that subsequent tests don't reuse
+    if "stack" in sys.modules:
+        del sys.modules["stack"]
+
+
+def test_synthesize_stack() -> None:
+    import stack
+
+    app = cdk.App()
+
+    project_name = "test-project"
+    dep_name = "test-deployment"
+    mod_name = "test-module"
+    app_prefix = f"{project_name}-{dep_name}-{mod_name}"
+
+    ecr_repo_name = "repo"
+
+    stack = stack.MlflowImagePublishingStack(
+        scope=app,
+        id=app_prefix,
+        app_prefix=app_prefix,
+        ecr_repo_name=ecr_repo_name,
+        env=cdk.Environment(
+            account=os.environ["CDK_DEFAULT_ACCOUNT"],
+            region=os.environ["CDK_DEFAULT_REGION"],
+        ),
+    )
+
+    template = Template.from_stack(stack)
+    template.resource_count_is("Custom::CDKBucketDeployment", 1)