+++ CHANGELOG.md
@@ -11,11 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### **Added**
 
+- added `mlflow-image` and `mlflow-fargate` modules
 - added `sagemaker-studio` module
 - added `sagemaker-endpoint` module
 
 ### **Changed**
 
 ### **Removed**
-
-=======